requirements.txt
build==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
Deprecated==1.2.13
-e git+ssh://git@github.com/UCLM-ESI/gh_org_invite.git@3bfd7f66ce3f4bbcb2c929931c765761454f278a#egg=gh_org_invite
idna==3.3
packaging==21.0
pep517==0.12.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
requests==2.26.0
six==1.16.0
tomli==1.2.1
urllib3==1.26.7
wrapt==1.13.2
